Fix some compiler warnings
Summary: There are several logic mistakes in the current code. Compiler warns about them and this patch tries to fix tnem.

Test Plan: Compiles and installs, XMPP part was not tested.

Reviewers: #kde_games, aacid

Reviewed By: aacid

Subscribers: nhirsl, aacid, chehrlic, kde-games-devel

Differential Revision: https://phabricator.kde.org/D25199

+++ ksirkskineditor/onu.cpp
@@ -56,7 +56,7 @@ ONU::ONU(const QString& configDir, QObject *parent)
 
   qCDebug(KSIRKSKINEDITOR_LOG) << "ONU constructor: " << m_configFileName;
   m_font.family = QStringLiteral("URW Chancery L");
-  m_font.size = (uint)(13);
+  m_font.size = 13;
   m_font.weight = QFont::Bold;
   m_font.italic = true;
   m_font.foregroundColor = QStringLiteral("black");
@@ -199,7 +199,7 @@ ONU::ONU(const QString& configDir, QObject *parent)
   qCDebug(KSIRKSKINEDITOR_LOG) << "Loading font data";
   KConfigGroup fontgroup = config.group("font");
   m_font.family = fontgroup.readEntry("family","URW Chancery L");
-  m_font.size = fontgroup.readEntry("size",(uint)(13));
+  m_font.size = fontgroup.readEntry("size", 13);
   QString w = fontgroup.readEntry("weight", "bold");;
   if (w == QLatin1String("normal"))
   {
